dom/base/test/test_audioNotificationNavigationWebAudio.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for audio controller in windows, when using the Web Audio API</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<pre id="test">
</pre>
<iframe></iframe>

<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

var expectedNotification = null;
var iframe = null;

var observer = {
  observe: function(subject, topic, data) {
    is(topic, "audio-playback", "audio-playback received");
    var expected = expectedNotification.shift();
    is(data, expected, `"${data}" is the right notification`);
    if (expectedNotification.length == 0) {
      runTest();
    }
  }
};

var observerService = SpecialPowers.Cc["@mozilla.org/observer-service;1"]
                                   .getService(SpecialPowers.Ci.nsIObserverService);

var tests = [
  function() {
    iframe = document.querySelector("iframe");
    observerService.addObserver(observer, "audio-playback");
    ok(true, "Observer set");
    runTest();
  },

  function() {
    info("Load iframe");
    expectedNotification = ["active"];
    iframe.src = "file_webAudioAudible.html";
  },

  function() {
    info("Reload iframe");
    // As reloading frame would stop previous audio playing in frame, we would
    // receive "inactive-pause" first.
    expectedNotification = ["inactive-pause", "active"];
    iframe.src = "file_webAudioAudible.html";
  },
  function() {
    info("Cleaning up iframe");
    expectedNotification = ["inactive-pause"];
    iframe.src = null;
  }
];

function runTest() {
  if (!tests.length) {
    observerService.removeObserver(observer, "audio-playback");
    ok(true, "Observer removed");
    SimpleTest.finish();
    return;
  }

  var test = tests.shift();
  test();
}

onload = runTest;

</script>
</body>
</html>
